create utils class for validation tests
This moves common functions out of one of the test classes into
a utils class, so that test classes don't have to import other test
classes.
